@@ -2,25 +2,25 @@
 	"name": "vite_app",
 	"version": "0.0.0",
 	"devDependencies": {
-		"@sveltejs/vite-plugin-svelte": "^3.0.2",
-		"@tsconfig/svelte": "^5.0.2",
-		"autoprefixer": "^10.4.19",
-		"eslint": "^8.57.0",
-		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
-		"postcss": "^8.4.38",
-		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
-		"prettier-plugin-tailwindcss": "^0.5.13",
-		"svelte": "^4.2.12",
-		"svelte-check": "^3.6.7",
-		"tailwindcss": "^3.4.4",
-		"tslib": "^2.6.2",
-		"typescript": "^5.4.3",
-		"typescript-eslint": "^7.4.0",
-		"vite": "^5.2.0",
+		"@sveltejs/vite-plugin-svelte": "^6.2.1",
+		"@tsconfig/svelte": "^5.0.5",
+		"autoprefixer": "^10.4.21",
+		"eslint": "^8.57.1",
+		"eslint-config-prettier": "^9.1.2",
+		"eslint-plugin-svelte": "^2.46.1",
+		"postcss": "^8.5.6",
+		"prettier": "^3.6.2",
+		"prettier-plugin-svelte": "^3.4.0",
+		"prettier-plugin-tailwindcss": "^0.5.14",
+		"svelte": "^5.39.6",
+		"svelte-check": "^4.3.2",
+		"tailwindcss": "^3.4.17",
+		"tslib": "^2.8.1",
+		"typescript": "^5.9.2",
+		"typescript-eslint": "^7.18.0",
+		"vite": "^7.1.7",
 		"vite-tsconfig-paths": "^4.3.2",
-		"vitest": "^1.2.0"
+		"vitest": "^3.2.4"
 	},
 	"private": true,
 	"scripts": {
@@ -35,5 +35,9 @@
 		"lint": "prettier --check . && eslint .",
 		"format": "prettier --write ."
 	},
-	"type": "module"
+	"type": "module",
+	"dependencies": {
+		"@typescript-eslint/eslint-plugin": "^8.44.1",
+		"@typescript-eslint/parser": "^8.44.1"
+	}
 }
