@@ -1,12 +1,12 @@
 <script lang="ts">
 	import AccountDropdownContent from '$lib/Components/AccountDropdownContent.svelte';
 	import Dropdown from '$lib/Components/CommonUI/Dropdown.svelte';
-	import SettingsModal from '$lib/Components/SettingsModal.svelte';
+	import SettingsModal from './Modals/SettingsModal.svelte';
 	import { AuthService } from '$lib/Services/AuthService';
 	import { BoltService } from '$lib/Services/BoltService';
 	import { bolt } from '$lib/State/Bolt';
 	import { GlobalState } from '$lib/State/GlobalState';
-	import { Game } from '$lib/Util/interfaces';
+	import { Game } from '$lib/Util/Interfaces';
 
 	const { config, sessions } = GlobalState;
 	let settingsModal: SettingsModal;
@@ -22,7 +22,7 @@
 			class="mx-1 w-20 rounded-lg border-2 border-blue-500 p-2 duration-200 hover:opacity-75"
 			class:text-black={$config.selected.game === Game.rs3}
 			class:bg-blue-500={$config.selected.game === Game.rs3}
-			on:click={() => {
+			onclick={() => {
 				$config.selected.game = Game.rs3;
 			}}
 		>
@@ -32,7 +32,7 @@
 			class="mx-1 w-20 rounded-lg border-2 border-blue-500 bg-blue-500 p-2 text-black duration-200 hover:opacity-75"
 			class:text-black={$config.selected.game === Game.osrs}
 			class:bg-blue-500={$config.selected.game === Game.osrs}
-			on:click={() => {
+			onclick={() => {
 				$config.selected.game = Game.osrs;
 			}}
 		>
@@ -42,13 +42,13 @@
 	<div class="m-2 ml-auto flex gap-2">
 		<button
 			class="h-10 w-10 rounded-full bg-blue-500 text-center duration-200 hover:rotate-45 hover:opacity-75"
-			on:click={() => ($config.use_dark_theme = !$config.use_dark_theme)}
+			onclick={() => ($config.use_dark_theme = !$config.use_dark_theme)}
 		>
 			<img src="svgs/lightbulb-solid.svg" class="m-auto h-6 w-6" alt="Change Theme" />
 		</button>
 		<button
 			class="h-10 w-10 rounded-full bg-blue-500 text-center duration-200 hover:rotate-45 hover:opacity-75"
-			on:click={() => settingsModal.open()}
+			onclick={() => settingsModal.open()}
 		>
 			<img src="svgs/gear-solid.svg" class="m-auto h-6 w-6" alt="Settings" />
 		</button>
@@ -61,14 +61,16 @@
 					{selectedSession?.user.displayName ?? 'No user selected'}
 				</button>
 
-				<div slot="content" class="w-40">
-					<AccountDropdownContent />
-				</div>
+				{#snippet content()}
+					<div class="w-40">
+						<AccountDropdownContent />
+					</div>
+				{/snippet}
 			</Dropdown>
 		{:else}
 			<button
 				class="h-11 w-48 rounded-lg border-2 border-slate-300 bg-inherit p-2 text-center font-bold text-black duration-200 hover:opacity-75 dark:border-slate-800 dark:text-slate-50"
-				on:click={() => {
+				onclick={() => {
 					const { origin, redirect, clientid } = bolt.env;
 					AuthService.openLoginWindow(origin, redirect, clientid);
 				}}
