@@ -1,4 +1,4 @@
-import type { GameClient } from '$lib/Util/interfaces';
+import type { GameClient } from '$lib/Util/Interfaces';
 import { writable, type Writable } from 'svelte/store';
 
 // TODO: remove clientList
